Rename base classes and remove model_string attr
