+++ src/floris/simulation/__init__.py
@@ -33,8 +33,7 @@ isort:skip_file
 # that should be included in the simulation package.
 # Since some of these depend on each other, the order
 # that they are listed here does matter.
-from .base_class import BaseClass
-from .base_class import BaseModel
+from .base import BaseClass, BaseModel
 from .turbine import Turbine, Ct, power, axial_induction, average_velocity
 from .farm import Farm
 from .grid import Grid, TurbineGrid  # , FlowFieldGrid
